+++ src/main/java/Girinoscope.java
@@ -1,3 +1,4 @@
+
 import org.hihan.girinoscope.ui.UI;
 
 /**